check if file exists

+++ audm_scraper.py
@@ -114,40 +114,46 @@ def main():
 
             os.makedirs(publications_dir + "/" + article["short_name"], exist_ok=True)
             p = audm.paragraphs([article["object_id"]])
-            # Articles are split up by paragraph and there can be quite a few. Although they are numbered and timestamped, makes more sense to join them
-            with alive_bar(len(p), force_tty=True) as filebar:
-                for f in p:
-                    file = audm.get_file(f["audio_filename"])
-                    filename = publications_dir + "/" + article["short_name"] + "/" + f["audio_filename"]
-                    with open(filename, "wb") as fz:
-                        fz.write(file.content)
-                        files.append({"filename": filename, "index": f["index"]})
-                    filebar()
-            fo = sorted(files, key = lambda i: i['index'])
-            # Temporary file to enable ffmpeg to demux and concat the m4a files
-            tempfile = os.path.join(publications_dir + "/" + article["short_name"], "templist.txt").replace("'", "'\\''")
 
-            eventual_outfile = os.path.join(publications_dir, article["short_name"] + "-" + article["author_name"] + "-" + article["pub_date"] + ".m4a").replace("'", "'\\''")
+            eventual_outfile = os.path.join(publications_dir,
+                                            article["short_name"] + "-" + article["author_name"] + "-" + article[
+                                                "pub_date"] + ".m4a").replace("'", "'\\''")
+            if not os.path.exists(eventual_outfile):
 
-            with open(tempfile, "a") as listf:
+                # Articles are split up by paragraph and there can be quite a few. Although they are numbered and timestamped, makes more sense to join them
+                with alive_bar(len(p), force_tty=True) as filebar:
+                    for f in p:
+                        file = audm.get_file(f["audio_filename"])
+                        filename = publications_dir + "/" + article["short_name"] + "/" + f["audio_filename"]
+                        with open(filename, "wb") as fz:
+                            fz.write(file.content)
+                            files.append({"filename": filename, "index": f["index"]})
+                        filebar()
+                fo = sorted(files, key = lambda i: i['index'])
+                # Temporary file to enable ffmpeg to demux and concat the m4a files
+                tempfile = os.path.join(publications_dir + "/" + article["short_name"], "templist.txt").replace("'", "'\\''")
 
-                for fn in fo:
-                    sanitized_file = fn["filename"].replace("'", "'\\''")
-                    listf.write("file '" + sanitized_file + "'\n")
 
-            concat_command = f"ffmpeg -nostats -loglevel 0 -y -f concat -safe 0 -i \"{tempfile}\" -c copy \"{eventual_outfile}\""
-            os.system(concat_command)
-            # Tagging
-            audio = taglib.File(eventual_outfile)
-            audio.tags["PERFORMER"] = article["narrator_name"]
-            audio.tags["ARTIST"] = article["author_name"]
-            audio.tags["TITLE"] = article["title"]
-            audio.tags["ALBUM"] = article["publication_name"]
-            audio.tags["DATE"] = article["pub_date"]
-            audio.tags["DESCRIPTION"] = article["desc"]
-            audio.save()
-            # Cleanup
-            shutil.rmtree(publications_dir + "/" + article["short_name"])
+
+                with open(tempfile, "a") as listf:
+
+                    for fn in fo:
+                        sanitized_file = fn["filename"].replace("'", "'\\''")
+                        listf.write("file '" + sanitized_file + "'\n")
+
+                concat_command = f"ffmpeg -nostats -loglevel 0 -y -f concat -safe 0 -i \"{tempfile}\" -c copy \"{eventual_outfile}\""
+                os.system(concat_command)
+                # Tagging
+                audio = taglib.File(eventual_outfile)
+                audio.tags["PERFORMER"] = article["narrator_name"]
+                audio.tags["ARTIST"] = article["author_name"]
+                audio.tags["TITLE"] = article["title"]
+                audio.tags["ALBUM"] = article["publication_name"]
+                audio.tags["DATE"] = article["pub_date"]
+                audio.tags["DESCRIPTION"] = article["desc"]
+                audio.save()
+                # Cleanup
+                shutil.rmtree(publications_dir + "/" + article["short_name"])
 
 
 if __name__ == '__main__':
